requirements.txt
gitpython
matplotlib
numpy>=1.18.2
pandas
